Late commit: fixed all issues, tests pass

+++ src/resources/boards/board.router.js
@@ -4,41 +4,40 @@ const boardsService = require('./board.service');
 
 const { endpoints } = require('../../config/endpoint.config');
 const StatusCodes = require('http-status-codes');
-const { NOT_FOUND_ERROR, asyncHandler } = require('../../error');
+const { NotFoundError, asyncHandler, errMessage } = require('../../error');
 
-const table_name = 'Boards';
-const entity = 'Board';
+const ENTITY = 'board';
 
 router
   .route(endpoints.root)
-  .get(async (req, res) => {
-    try {
+  .get(
+    asyncHandler(async (req, res) => {
       const boards = await boardsService.getAllBoards();
+      if (!boards) throw new NotFoundError(errMessage.notFound(ENTITY));
+
       await res
         .status(StatusCodes.OK)
         .json(boards.map(board => Board.toResponse(board)));
-    } catch (err) {
-      NOT_FOUND_ERROR(res, table_name);
-    }
-  })
+    })
+  )
   .post(
     asyncHandler(async (req, res) => {
-      const board = new Board.entity(req.body);
-      const newBoard = await boardsService.createBoard(board);
-      res.status(StatusCodes.OK).send(Board.toResponse(newBoard));
+      const board = await boardsService.createBoard(new Board.entity(req.body));
+      res.status(StatusCodes.OK).send(Board.toResponse(board));
     })
   );
 
 router
   .route(endpoints.id)
-  .get(async (req, res) => {
-    try {
+  .get(
+    asyncHandler(async (req, res) => {
       const board = await boardsService.getBoardById(req.params.id);
+      if (!board) {
+        throw new NotFoundError(errMessage.notFoundParams(ENTITY, req.params));
+      }
       await res.status(StatusCodes.OK).send(Board.toResponse(board));
-    } catch (err) {
-      NOT_FOUND_ERROR(res, entity, req.params);
-    }
-  })
+    })
+  )
   .put(
     asyncHandler(async (req, res) => {
       const updatedBoard = await boardsService.updateBoard(
@@ -48,14 +47,15 @@ router
       res.status(StatusCodes.OK).send(Board.toResponse(updatedBoard));
     })
   )
-  .delete(async (req, res) => {
-    try {
+  .delete(
+    asyncHandler(async (req, res) => {
+      const board = await boardsService.getBoardById(req.params.id);
+      if (!board) {
+        throw new NotFoundError(errMessage.notFoundParams(ENTITY, req.params));
+      }
       await boardsService.deleteBoard(req.params.id);
       res.sendStatus(StatusCodes.NO_CONTENT);
-    } catch (err) {
-      console.log('BOARD DELETE...', err);
-      NOT_FOUND_ERROR(res, entity);
-    }
-  });
+    })
+  );
 
 module.exports = router;
